Complete revamp of error handling for commands.
The largest amount of work was done around making sure that writes
of plenary templates can be reversed.  The PlenaryCollection object
was enhanced to allow it to act as a generic wrapper for dealing
with errors in the simple cases.

Additionally all writes now force a refresh of the plenary's
database object.  This now requires that the session be flushed
before attempting a write.

Plenary.read() had a bug that checked the wrong location that was
fixed as part of this work.  When Plenary.read() was changed to use
the pathname() helper its use of read_file() was broken.  The
pathname() method returns a full path whereas read_file() expects the
path to be broken into a directory and a file name.

Also removed (fixed) the writing of plenary templates for machine_type
aurora_node.

+++ lib/python2.5/aquilon/server/templates/base.py
@@ -33,7 +33,9 @@ import os
 from threading import Lock
 
 from twisted.python import log
-from aquilon.exceptions_ import InternalError
+from sqlalchemy.orm.session import object_session
+
+from aquilon.exceptions_ import InternalError, IncompleteError
 from aquilon.config import Config
 from aquilon.server.processes import write_file, read_file, remove_file
 
@@ -59,8 +61,9 @@ def compileRelease():
 
 
 class Plenary(object):
-    def __init__(self):
+    def __init__(self, dbobj=None):
         self.config = Config()
+        self.dbobj = dbobj
         self.template_type = 'structure'
         self.plenary_template = None
         self.plenary_core = None
@@ -68,6 +71,9 @@ class Plenary(object):
         # The following attributes are for stash/restore_stash
         self.old_content = None
         self.old_mtime = None
+        self.stashed = False
+        self.removed = False
+        self.changed = False
 
     def pathname(self):
         return os.path.join(self.dir, self.plenary_template + ".tpl")
@@ -84,23 +90,31 @@ class Plenary(object):
         pass
 
     def write(self, dir=None, user=None, locked=False, content=None):
-        """
-        write out the template.
+        """Write out the template.
 
-        Writes out the template. If the content is unchanged, then
-        the file will not be modified (preserving the mtime).
-        Returns the number of files that were written
+        Note that session.flush() MUST be called before writing a
+        changed sqlalchemy object or the embedded call to refresh
+        will revert the object back to its old state.
+
+        If the content is unchanged, then the file will not be modified
+        (preserving the mtime).
+
+        Returns the number of files that were written.
+
+        If locked is True then it is assumed that error handling happens
+        higher in the call stack.
+
         """
 
         if dir is not None:
             self.dir = dir
         # user is simply left for compatibility: it's no longer used
-        if (hasattr(self, "machine_type") and
-                getattr(self, "machine_type") == 'aurora_node'):
-            # Don't bother writing plenary files for dummy aurora hardware.
-            return 0
 
         if content is None:
+            # This potentially calls refresh several times for objects that
+            # have multiple plenaries, like service and cluster...
+            self.refresh()
+
             lines = []
             type = self.template_type
             if type is not None and type is not "":
@@ -112,21 +126,28 @@ class Plenary(object):
 
         plenary_path = os.path.join(self.dir, self.plenary_core)
         plenary_file = self.pathname()
-        # optimise out the write (leaving the mtime good for make)
-        # if nothing is actually changed
-        if os.path.exists(plenary_file):
-            old = self.read()
-            if (old == content):
-                return 0
-            
-        if (not locked):
-            compileLock()
+
         try:
+            if not locked:
+                compileLock()
+            self.stash()
+            if self.old_content == content and \
+               not self.removed and not self.changed:
+                # optimise out the write (leaving the mtime good for ant)
+                # if nothing is actually changed
+                return 0
             if not os.path.exists(plenary_path):
                 os.makedirs(plenary_path)
             write_file(plenary_file, content)
+            self.removed = False
+            if self.old_content != content:
+                self.changed = True
+        except Exception, e:
+            if not locked:
+                self.restore_stash()
+            raise e
         finally:
-            if (not locked):
+            if not locked:
                 compileRelease()
 
         return 1
@@ -134,7 +155,8 @@ class Plenary(object):
     def read(self, dir=None):
         if dir is not None:
             self.dir = dir
-        return read_file(self.dir, self.pathname())
+        # FIXME: Dupes some logic from pathname()
+        return read_file(self.dir, self.plenary_template + ".tpl")
 
     def remove(self, dir=None, locked=False):
         """
@@ -147,39 +169,81 @@ class Plenary(object):
         try:
             if (not locked):
                 compileLock()
+            self.stash()
             remove_file(self.pathname())
+            self.removed = True
+        # Most of the error handling routines would restore_stash...
+        # but there's no need here if the remove failed. :)
         finally:
             if (not locked):
                 compileRelease()
         return
 
-    def cleanup(self, name, domain, locked=False):
+    def cleanup(self, domain, locked=False):
         """
         remove all files related to an object template including
         any intermediate build files
         """
-        self.remove(None, locked)
-        if self.template_type == "object":
-            qdir = self.config.get("broker", "quattordir")
-            xmlfile = os.path.join(qdir, "build", "xml", domain, name+".xml")
-            remove_file(xmlfile)
-            depfile = os.path.join(qdir, "build", "xml", domain, name+".xml.dep")
-            remove_file(depfile)
+        try:
+            if not locked:
+                compileLock()
+            self.remove(locked=True)
+            if self.template_type == "object" and hasattr(self, 'name'):
+                qdir = self.config.get("broker", "quattordir")
+                xmlfile = os.path.join(qdir, "build", "xml", domain,
+                                       self.name + ".xml")
+                remove_file(xmlfile)
+                depfile = os.path.join(qdir, "build", "xml", domain,
+                                       self.name + ".xml.dep")
+                remove_file(depfile)
+        except:
+            if not locked:
+                self.restore_stash()
+            raise
+        finally:
+            if not locked:
+                compileRelease()
+
+    def refresh(self):
+        """Refresh any stored database objects.
+
+        Note that the session MUST be flushed before calling a refresh
+        if the object's state has changed or it will be reverted.
+
+        """
+        # This could be enhanced to deal with dbobj being a list...
+        if hasattr(self, 'dbobj') and self.dbobj:
+            session = object_session(self.dbobj)
+            session.refresh(self.dbobj)
 
     def stash(self):
+        """Record the state of the plenary to make restoration possible.
+
+        This should only be called while holding a compileLock.
+
         """
-        record the state of the plenary, in order to later restore this state
-        """
+        if self.stashed:
+            return
         try:
             self.old_content = self.read()
             self.old_mtime = os.stat(self.pathname()).st_atime
         except IOError, e:
             self.old_content = None
+        self.stashed = True
 
     def restore_stash(self):
+        """Restore previous state of plenary.
+
+        This should only be called while holding a compileLock.
+
         """
-        restore previous state of plenary
-        """
+        if not self.stashed:
+            log.msg("Attempt to restore plenary '%s' "
+                    "without having saved state." % self.pathname())
+            return
+        # Should this optimization be in use?
+        # if not self.changed and not self.removed:
+        #    return
         if (self.old_content is None):
             self.remove(locked=True)
         else:
@@ -187,12 +251,25 @@ class Plenary(object):
             atime = os.stat(self.pathname()).st_atime
             os.utime(self.pathname(), (atime, self.old_mtime))
 
-class PlenaryCollection():
+
+class PlenaryCollection(object):
     """
     A collection of plenary templates, presented behind a single
-    facade to make them appear as one template to the caller. To use,
-    subclass from this and append the real template objects into
+    facade to make them appear as one template to the caller.
+
+    One use is for objects that logically have multiple plenary files
+    to subclass from this and append the real template objects into
     self.plenaries.
+
+    Another is for commands to use this object directly for its
+    convenience methods around writing and rolling back plenary
+    templates.
+
+    This object cannot handle cases like a plenary file that
+    changes location, but it could handle storing a group of plenaries
+    that need to be removed and a second collection could handle another
+    group that needs to be written.
+
     """
     def __init__(self):
         self.plenaries = []
@@ -206,20 +283,71 @@ class PlenaryCollection():
             plen.restore_stash()
 
     def write(self, dir=None, user=None, locked=False, content=None):
+        # If locked is True, assume error handling happens higher
+        # in the stack.
         total = 0
-        for plen in self.plenaries:
-            total += plen.write(dir, user, locked, content)
+        try:
+            if not locked:
+                compileLock()
+            # Pre-stash all plenaries before attempting to write any
+            # of them.  This way if an error occurs all can go through
+            # the same restore logic.
+            self.stash()
+            for plen in self.plenaries:
+                # IncompleteError is almost pointless in this context, but
+                # it has the nice side effect of not updating the total.
+                try:
+                    total += plen.write(dir=dir, user=user, locked=True,
+                                        content=content)
+                except IncompleteError, e:
+                    pass
+        except:
+            if not locked:
+                self.restore_stash()
+            raise
+        finally:
+            if not locked:
+                compileRelease()
         return total
 
     def remove(self, dir=None, locked=False):
-        for plen in self.plenaries:
-            plen.remove(dir, locked)
+        try:
+            if not locked:
+                compileLock()
+            self.stash()
+            for plen in self.plenaries:
+                plen.remove(dir, locked=True)
+        except:
+            if not locked:
+                self.restore_stash()
+            raise
+        finally:
+            if not locked:
+                compileRelease()
 
-    def cleanup(self, name, domain, locked=False):
-        for plen in self.plenaries:
-            plen.cleanup(name, domain, locked)
+    def cleanup(self, domain, locked=False):
+        try:
+            if not locked:
+                compileLock()
+            self.stash()
+            for plen in self.plenaries:
+                plen.cleanup(domain, locked=True)
+        except:
+            if not locked:
+                self.restore_stash()
+            raise
+        finally:
+            if not locked:
+                compileRelease()
 
     def read(self):
         # This should never be called, but we put it here
         # just in-case, since the base-class method is inappropriate.
         raise InternalError
+
+    def append(self, plenary):
+        self.plenaries.append(plenary)
+
+    def refresh(self):
+        for plen in self.plenaries:
+            plen.refresh()
